Feature: reportcount.
Initial step to run xping in a non-interactive or unattended fashion.
Number of probes are specified using "-c num" and nothing is displayed
during probing and when (interval+1)*num have elapsed the program stops
and shows the result on the screen in the usual layout (but with out
all the control codes).

This feature is similar to mtr's --report and --report-cycles.